Second part
Assisted by Cursor
