+++ src/platform/test/functional/apps/discover/query_mode/_default_query_mode.ts
@@ -6,60 +6,9 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-import expect from '@kbn/expect';
 import type { FtrProviderContext } from '../ftr_provider_context';
+import { defineDefaultQueryModeTests } from './_shared_default_query_mode_tests';
 
-export default function ({ getPageObjects, getService }: FtrProviderContext) {
-  const { discover, common, unifiedSearch } = getPageObjects([
-    'discover',
-    'common',
-    'unifiedSearch',
-  ]);
-
-  describe('Default query mode', () => {
-    afterEach(async () => {
-      await discover.resetQueryMode();
-    });
-
-    describe('when there is no default query mode set', () => {
-      it('should open Discover in classic mode', async () => {
-        // Validate that no default query mode is set
-        await common.navigateToApp('discover');
-        const queryMode = await discover.getQueryMode();
-        expect(queryMode).to.be(null);
-
-        // Go to discover and validate classic mode
-        await discover.isInClassicMode();
-      });
-    });
-
-    describe('when the user clicks ES|QL mode', () => {
-      it('should set the default mode to ES|QL', async () => {
-        // Go to discover and select ES|QL mode
-        await common.navigateToApp('discover');
-        await discover.selectTextBaseLang();
-        const queryMode = await discover.getQueryMode();
-        expect(queryMode).to.contain('esql');
-
-        // Reload the app and validate ES|QL mode is persisted
-        await common.navigateToApp('discover', { path: '' });
-        await discover.isInEsqlMode();
-      });
-    });
-
-    describe('when the user clicks classic', () => {
-      it('should set the default mode to classic', async () => {
-        // Go to discover and select classic mode
-        await common.navigateToApp('discover');
-        await discover.selectTextBaseLang();
-        await unifiedSearch.switchToDataViewMode();
-        const queryMode = await discover.getQueryMode();
-        expect(queryMode).to.contain('classic');
-
-        // Reload the app and validate classic mode is persisted
-        await common.navigateToApp('discover', { path: '' });
-        await discover.isInClassicMode();
-      });
-    });
-  });
+export default function (ctx: FtrProviderContext) {
+  defineDefaultQueryModeTests(ctx, { isEsqlDefault: false });
 }
